api folder renamed to api2 with updated imports by vscode
