+++ src/components/mui/SignIn.js
@@ -12,7 +12,7 @@ import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
 import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
-import ServerApi from '../../api/api';
+import ServerApi from '../../api2/api';
 import { useNavigate, useLocation } from "react-router-dom";
 function Copyright(props) {
   return (
@@ -44,7 +44,8 @@ export default function SignUp() {
       surname: data.get('lastname')
     };
     const res = await ServerApi.doPost('/users/signup', formdata)
-    navigate("/")
+    console.log('response: ' + JSON.stringify(res))
+    // navigate("/")
   };
 
   return (
